+++ orm/load.go
@@ -2,16 +2,18 @@ package orm
 
 import (
 	"context"
+	"database/sql"
 	"fmt"
 	"reflect"
 
+	"github.com/aarondl/opt"
 	"github.com/stephenafamo/bob"
+	"github.com/stephenafamo/bob/clause"
+	"github.com/stephenafamo/bob/expr"
+	"github.com/stephenafamo/bob/internal"
+	"github.com/stephenafamo/bob/mods"
 	"github.com/stephenafamo/scan"
 )
-
-type Preloadable interface {
-	Preload(name string, rel any) error
-}
 
 // Loadable is a constraint for types that can be loaded
 type Loadable interface {
@@ -181,4 +183,208 @@ func (a *AfterPreloader) Load(ctx context.Context, exec bob.Executor, _ any) err
 	}
 
 	return nil
+}
+
+type Preloadable interface {
+	Preload(name string, rel any) error
+}
+
+type PreloadRel[E bob.Expression] struct {
+	Name  string
+	Sides []PreloadSide[E]
+}
+
+type nameable[E bob.Expression] interface {
+	Name() E
+	Alias() string
+}
+
+type PreloadSide[E bob.Expression] struct {
+	From        nameable[E]
+	To          nameable[E]
+	FromColumns []string `yaml:"-"`
+	ToColumns   []string `yaml:"-"`
+
+	FromWhere []RelWhere `yaml:"from_where"`
+	ToWhere   []RelWhere `yaml:"to_where"`
+}
+
+type PreloadableQuery interface {
+	Loadable
+	AppendJoin(clause.Join)
+	AppendPreloadSelect(columns ...any)
+}
+
+func Preload[T Preloadable, Ts ~[]T, E bob.Expression, Q PreloadableQuery](rel PreloadRel[E], cols []string, opts ...PreloadOption[Q]) Preloader[Q] {
+	settings := NewPreloadSettings[T, Ts, Q](cols)
+	for _, o := range opts {
+		if o == nil {
+			continue
+		}
+		o.ModifyPreloadSettings(&settings)
+	}
+
+	return buildPreloader[T](func(parent string) (string, mods.QueryMods[Q]) {
+		if parent == "" {
+			parent = rel.Sides[0].From.Alias()
+		}
+
+		var alias string
+		var queryMods mods.QueryMods[Q]
+
+		for i, side := range rel.Sides {
+			alias = settings.Alias
+			if settings.Alias == "" {
+				alias = fmt.Sprintf("%s_%d", side.To, internal.RandInt())
+			}
+			on := make([]bob.Expression, 0, len(side.FromColumns)+len(side.FromWhere)+len(side.ToWhere))
+			for i, fromCol := range side.FromColumns {
+				toCol := side.ToColumns[i]
+				on = append(on, expr.OP(
+					"=",
+					expr.Quote(parent, fromCol),
+					expr.Quote(alias, toCol),
+				))
+			}
+			for _, from := range side.FromWhere {
+				on = append(on, expr.OP(
+					"=",
+					expr.Quote(parent, from.Column),
+					expr.Raw(from.SQLValue),
+				))
+			}
+			for _, to := range side.ToWhere {
+				on = append(on, expr.OP(
+					"=",
+					expr.Quote(parent, to.Column),
+					expr.Raw(to.SQLValue),
+				))
+			}
+
+			if len(settings.Mods) > i {
+				for _, additional := range settings.Mods[i] {
+					on = append(on, additional(parent, alias)...)
+				}
+			}
+
+			queryMods = append(queryMods, mods.Join[Q](clause.Join{
+				Type: clause.LeftJoin,
+				To: clause.TableRef{
+					Expression: side.To.Name(),
+					Alias:      alias,
+				},
+				On: on,
+			}))
+
+			// so the condition on the next "side" will be on the right table
+			parent = alias
+		}
+
+		queryMods = append(queryMods, mods.Preload[Q]{
+			NewColumns(settings.Columns...).WithParent(alias).WithPrefix(alias + "."),
+		})
+		return alias, queryMods
+	}, rel.Name, settings)
+}
+
+func buildPreloader[T any, Q Loadable](f func(string) (string, mods.QueryMods[Q]), name string, opt PreloadSettings[Q]) Preloader[Q] {
+	return func(parent string) (bob.Mod[Q], scan.MapperMod, []bob.Loader) {
+		alias, queryMods := f(parent)
+		prefix := alias + "."
+
+		var mapperMods []scan.MapperMod
+		extraLoaders := []bob.Loader{opt.ExtraLoader}
+
+		for _, l := range opt.SubLoaders {
+			queryMod, mapperMod, extraLoader := l(alias)
+			if queryMod != nil {
+				queryMods = append(queryMods, queryMod)
+			}
+
+			if mapperMod != nil {
+				mapperMods = append(mapperMods, mapperMod)
+			}
+
+			if extraLoader != nil {
+				extraLoaders = append(extraLoaders, extraLoader...)
+			}
+		}
+
+		return queryMods, func(ctx context.Context, cols []string) (scan.BeforeFunc, scan.AfterMod) {
+			before, after := scan.StructMapper[T](
+				scan.WithStructTagPrefix(prefix),
+				scan.WithTypeConverter(typeConverter{}),
+				scan.WithRowValidator(rowValidator),
+				scan.WithMapperMods(mapperMods...),
+			)(ctx, cols)
+
+			return before, func(link, retrieved any) error {
+				loader, isLoader := retrieved.(Preloadable)
+				if !isLoader {
+					return fmt.Errorf("object %T cannot pre load", retrieved)
+				}
+
+				t, err := after(link)
+				if err != nil {
+					return err
+				}
+
+				if err = opt.ExtraLoader.Collect(t); err != nil {
+					return err
+				}
+
+				return loader.Preload(name, t)
+			}
+		}, extraLoaders
+	}
+}
+
+// the row is valid if at least one column is not null
+func rowValidator(_ []string, vals []reflect.Value) bool {
+	for _, v := range vals {
+		v, ok := v.Interface().(*wrapper)
+		if !ok {
+			return false
+		}
+
+		if !v.IsNull {
+			return true
+		}
+	}
+
+	return false
+}
+
+type wrapper struct {
+	IsNull bool
+	V      any
+}
+
+// Scan implements the sql.Scanner interface. If the wrapped type implements
+// sql.Scanner then it will call that.
+func (v *wrapper) Scan(value any) error {
+	if value == nil {
+		v.IsNull = true
+		return nil
+	}
+
+	if scanner, ok := v.V.(sql.Scanner); ok {
+		return scanner.Scan(value)
+	}
+
+	return opt.ConvertAssign(v.V, value)
+}
+
+type typeConverter struct{}
+
+func (typeConverter) TypeToDestination(typ reflect.Type) reflect.Value {
+	val := reflect.ValueOf(&wrapper{
+		V: reflect.New(typ).Interface(),
+	})
+
+	return val
+}
+
+func (typeConverter) ValueFromDestination(val reflect.Value) reflect.Value {
+	return val.Elem().FieldByName("V").Elem().Elem()
 }